+++ confirmation/confirmation.go
@@ -1,0 +1,7 @@
+package confirmation
+
+import "context"
+
+type ConfirmationAccessor interface {
+	DeleteUserConfirmations(ctx context.Context, userID string) error
+}