+++ confirmation/store/store.go
@@ -1,19 +1,17 @@
 package store
 
 import (
-	"context"
-
+	"github.com/tidepool-org/platform/confirmation"
 	"github.com/tidepool-org/platform/store"
 )
 
 type Store interface {
 	store.Store
 
-	NewConfirmationsSession() ConfirmationsSession
+	NewConfirmationSession() ConfirmationSession
 }
 
-type ConfirmationsSession interface {
+type ConfirmationSession interface {
 	store.Session
-
-	DestroyConfirmationsForUserByID(ctx context.Context, userID string) error
+	confirmation.ConfirmationAccessor
 }